+++ CHANGES.md
@@ -60,6 +60,8 @@
     Can be disabled by adding ``pat-inject-errorhandler.off`` to the URL's query string.
 -   core utils: Add ``jqToNode`` to return a DOM node if a jQuery node was passed.
 -   pat inject: Rebase URLs in pattern configuration attributes. This avoids URLs in pattern configuration to point to unreachable paths in the context where the result is injected into.
+-   pat forward: Add `delay` option for delaying the click action forwarding for a given number of milliseconds.
+-   pat forward: Add `self` as possible value for the `selector` option to trigger the event on itself.
 
 ### Technical